+++ pkg/iam/control_plane_template.go
@@ -28,6 +28,7 @@ const controlPlanePolicyTemplate = `{
         "autoscaling:DescribeAutoScalingInstances",
         "autoscaling:DescribeTags",
         "autoscaling:DescribeLaunchConfigurations",
+        "autoscaling:DescribeScalingActivities",
         "ec2:DescribeLaunchTemplateVersions"
       ],
       "Resource": "*",