+++ pkg/iam/nodes_template.go
@@ -19,6 +19,7 @@ const nodesTemplate = `{
         "autoscaling:DescribeAutoScalingInstances",
         "autoscaling:DescribeTags",
         "autoscaling:DescribeLaunchConfigurations",
+        "autoscaling:DescribeScalingActivities",
         "ec2:DescribeLaunchTemplateVersions"
       ],
       "Resource": "*",